+++ dart/test/dynamic_color_test.dart
@@ -43,58 +43,57 @@ final red = Hct.fromInt(0xFFFF0000);
 
 final _colors = <String, DynamicColor>{
   'background': MaterialDynamicColors.background,
-  'onBackground': MaterialDynamicColors.onBackground,
+  'on_background': MaterialDynamicColors.onBackground,
   'surface': MaterialDynamicColors.surface,
-  'surfaceDark': MaterialDynamicColors.surfaceDark,
-  'surfaceLight': MaterialDynamicColors.surfaceLight,
-  'surfaceSub2': MaterialDynamicColors.surfaceSub2,
-  'surfaceSub1': MaterialDynamicColors.surfaceSub1,
-  'surfaceContainer': MaterialDynamicColors.surfaceContainer,
-  'surfaceAdd1': MaterialDynamicColors.surfaceAdd1,
-  'surfaceAdd2': MaterialDynamicColors.surfaceAdd2,
-  'onSurface': MaterialDynamicColors.onSurface,
-  'surfaceVariant': MaterialDynamicColors.surfaceVariant,
-  'onSurfaceVariant': MaterialDynamicColors.onSurfaceVariant,
-  'surfaceInverse': MaterialDynamicColors.surfaceInverse,
-  'onSurfaceInverse': MaterialDynamicColors.onSurfaceInverse,
+  'surface_dim': MaterialDynamicColors.surfaceDim,
+  'surface_bright': MaterialDynamicColors.surfaceBright,
+  'surface_container_lowest': MaterialDynamicColors.surfaceContainerLowest,
+  'surface_container_low': MaterialDynamicColors.surfaceContainerLow,
+  'surface_container': MaterialDynamicColors.surfaceContainer,
+  'surface_container_high': MaterialDynamicColors.surfaceContainerHigh,
+  'surface_container_highest': MaterialDynamicColors.surfaceContainerHighest,
+  'on_surface': MaterialDynamicColors.onSurface,
+  'surface_variant': MaterialDynamicColors.surfaceVariant,
+  'on_surface_variant': MaterialDynamicColors.onSurfaceVariant,
+  'inverse_surface': MaterialDynamicColors.inverseSurface,
+  'inverse_on_surface': MaterialDynamicColors.inverseOnSurface,
   'outline': MaterialDynamicColors.outline,
-  'outlineVariant': MaterialDynamicColors.outlineVariant,
+  'outline_variant': MaterialDynamicColors.outlineVariant,
   'shadow': MaterialDynamicColors.shadow,
   'scrim': MaterialDynamicColors.scrim,
-  'surfaceTint': MaterialDynamicColors.surfaceTint,
+  'surface_tint': MaterialDynamicColors.surfaceTint,
   'primary': MaterialDynamicColors.primary,
-  'onPrimary': MaterialDynamicColors.onPrimary,
-  'primaryContainer': MaterialDynamicColors.primaryContainer,
-  'onPrimaryContainer': MaterialDynamicColors.onPrimaryContainer,
-  'primaryInverse': MaterialDynamicColors.primaryInverse,
-  'onPrimaryInverse': MaterialDynamicColors.onPrimaryInverse,
+  'on_primary': MaterialDynamicColors.onPrimary,
+  'primary_container': MaterialDynamicColors.primaryContainer,
+  'on_primary_container': MaterialDynamicColors.onPrimaryContainer,
+  'inverse_primary': MaterialDynamicColors.inversePrimary,
   'secondary': MaterialDynamicColors.secondary,
-  'onSecondary': MaterialDynamicColors.onSecondary,
-  'secondaryContainer': MaterialDynamicColors.secondaryContainer,
-  'onSecondaryContainer': MaterialDynamicColors.onSecondaryContainer,
+  'on_secondary': MaterialDynamicColors.onSecondary,
+  'secondary_container': MaterialDynamicColors.secondaryContainer,
+  'on_secondary_container': MaterialDynamicColors.onSecondaryContainer,
   'tertiary': MaterialDynamicColors.tertiary,
-  'onTertiary': MaterialDynamicColors.onTertiary,
-  'tertiaryContainer': MaterialDynamicColors.tertiaryContainer,
-  'onTertiaryContainer': MaterialDynamicColors.onTertiaryContainer,
+  'on_tertiary': MaterialDynamicColors.onTertiary,
+  'tertiary_container': MaterialDynamicColors.tertiaryContainer,
+  'on_tertiary_container': MaterialDynamicColors.onTertiaryContainer,
   'error': MaterialDynamicColors.error,
-  'onError': MaterialDynamicColors.onError,
-  'errorContainer': MaterialDynamicColors.errorContainer,
-  'onErrorContainer': MaterialDynamicColors.onErrorContainer,
+  'on_error': MaterialDynamicColors.onError,
+  'error_container': MaterialDynamicColors.errorContainer,
+  'on_error_container': MaterialDynamicColors.onErrorContainer,
 };
 
 final _textSurfacePairs = [
-  _Pair('onPrimary', 'primary'),
-  _Pair('onPrimaryContainer', 'primaryContainer'),
-  _Pair('onSecondary', 'secondary'),
-  _Pair('onSecondaryContainer', 'secondaryContainer'),
-  _Pair('onTertiary', 'tertiary'),
-  _Pair('onTertiaryContainer', 'tertiaryContainer'),
-  _Pair('onError', 'error'),
-  _Pair('onErrorContainer', 'errorContainer'),
-  _Pair('onBackground', 'background'),
-  _Pair('onSurfaceVariant', 'surfaceVariant'),
-  _Pair('onSurfaceInverse', 'surfaceInverse'),
-  _Pair('onPrimaryInverse', 'primaryInverse'),
+  _Pair('on_primary', 'primary'),
+  _Pair('on_primary_container', 'primary_container'),
+  _Pair('on_secondary', 'secondary'),
+  _Pair('on_secondary_container', 'secondary_container'),
+  _Pair('on_tertiary', 'tertiary'),
+  _Pair('on_tertiary_container', 'tertiary_container'),
+  _Pair('on_error', 'error'),
+  _Pair('on_error_container', 'error_container'),
+  _Pair('on_background', 'background'),
+  _Pair('on_surface_variant', 'surface_bright'),
+  _Pair('on_surface_variant', 'surface_dim'),
+  _Pair('inverse_on_surface', 'inverse_surface'),
 ];
 
 void main() {
@@ -105,7 +104,7 @@ void main() {
         isDark: false,
         contrastLevel: 0.5,
       )),
-      equals(0xFFFFE5E1),
+      equals(0xFFFFFFFF),
     );
     expect(
       MaterialDynamicColors.onSecondaryContainer.getArgb(SchemeContent(
@@ -113,7 +112,7 @@ void main() {
         isDark: false,
         contrastLevel: 0.5,
       )),
-      equals(0xFFFFFCFF),
+      equals(0xFFFFFFFF),
     );
     expect(
       MaterialDynamicColors.onTertiaryContainer.getArgb(SchemeContent(
@@ -121,21 +120,21 @@ void main() {
         isDark: true,
         contrastLevel: -0.5,
       )),
-      equals(0xFF616600),
+      equals(0xffbac040),
     );
     expect(
-      MaterialDynamicColors.surfaceInverse.getArgb(SchemeContent(
+      MaterialDynamicColors.inverseSurface.getArgb(SchemeContent(
           sourceColorHct: Hct.fromInt(0xFF0000FF),
           isDark: false,
           contrastLevel: 0.0)),
-      equals(0xFF464652),
+      equals(0xFF2F2F3B),
     );
     expect(
-      MaterialDynamicColors.primaryInverse.getArgb(SchemeContent(
+      MaterialDynamicColors.inversePrimary.getArgb(SchemeContent(
           sourceColorHct: Hct.fromInt(0xFFFF0000),
           isDark: false,
           contrastLevel: -0.5)),
-      equals(0xFFFF8C7A),
+      equals(0xffff422f),
     );
     expect(
       MaterialDynamicColors.outlineVariant.getArgb(SchemeContent(
@@ -199,4 +198,176 @@ void main() {
       }
     }
   }
+
+  // Tests for fixed colors.
+  test('fixed colors in non-monochrome schemes', () {
+    final scheme = SchemeTonalSpot(
+      sourceColorHct: Hct.fromInt(0xFFFF0000),
+      isDark: true,
+      contrastLevel: 0.0,
+    );
+
+    expect(
+      MaterialDynamicColors.primaryFixed.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.primaryFixedDim.getHct(scheme).tone,
+      closeTo(80.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixed.getHct(scheme).tone,
+      closeTo(10.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixedVariant.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixed.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixedDim.getHct(scheme).tone,
+      closeTo(80.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixed.getHct(scheme).tone,
+      closeTo(10.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixedVariant.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixed.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixedDim.getHct(scheme).tone,
+      closeTo(80.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixed.getHct(scheme).tone,
+      closeTo(10.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixedVariant.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+  });
+
+  test('fixed colors in light monochrome schemes', () {
+    final scheme = SchemeMonochrome(
+      sourceColorHct: Hct.fromInt(0xFFFF0000),
+      isDark: false,
+      contrastLevel: 0.0,
+    );
+
+    expect(
+      MaterialDynamicColors.primaryFixed.getHct(scheme).tone,
+      closeTo(40.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.primaryFixedDim.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixed.getHct(scheme).tone,
+      closeTo(100.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixedVariant.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixed.getHct(scheme).tone,
+      closeTo(80.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixedDim.getHct(scheme).tone,
+      closeTo(70.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixed.getHct(scheme).tone,
+      closeTo(10.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixedVariant.getHct(scheme).tone,
+      closeTo(25.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixed.getHct(scheme).tone,
+      closeTo(40.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixedDim.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixed.getHct(scheme).tone,
+      closeTo(100.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixedVariant.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+  });
+
+  test('fixed colors in dark monochrome schemes', () {
+    final scheme = SchemeMonochrome(
+      sourceColorHct: Hct.fromInt(0xFFFF0000),
+      isDark: true,
+      contrastLevel: 0.0,
+    );
+
+    expect(
+      MaterialDynamicColors.primaryFixed.getHct(scheme).tone,
+      closeTo(40.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.primaryFixedDim.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixed.getHct(scheme).tone,
+      closeTo(100.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onPrimaryFixedVariant.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixed.getHct(scheme).tone,
+      closeTo(80.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.secondaryFixedDim.getHct(scheme).tone,
+      closeTo(70.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixed.getHct(scheme).tone,
+      closeTo(10.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onSecondaryFixedVariant.getHct(scheme).tone,
+      closeTo(25.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixed.getHct(scheme).tone,
+      closeTo(40.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.tertiaryFixedDim.getHct(scheme).tone,
+      closeTo(30.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixed.getHct(scheme).tone,
+      closeTo(100.0, 1.0),
+    );
+    expect(
+      MaterialDynamicColors.onTertiaryFixedVariant.getHct(scheme).tone,
+      closeTo(90.0, 1.0),
+    );
+  });
 }